requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pillow==10.3.0
requests==2.31.0
typing==3.7.4.3
typing_extensions==4.11.0
urllib3==2.2.1
